+++ packages/client/src/re-exports.ts
@@ -1,15 +1,17 @@
 export type {
   Codec,
   SS58String,
+  SS58AddressInfo,
   HexString,
   GetEnum,
   ResultPayload,
+  EnumVariant,
 } from "@polkadot-api/substrate-bindings"
 export {
   AccountId,
   Binary,
   FixedSizeBinary,
   Enum,
-  type EnumVariant,
   _Enum,
+  getSs58AddressInfo,
 } from "@polkadot-api/substrate-bindings"